+++ CHANGELOG.md
@@ -5,6 +5,7 @@ API deprecations:
 Known issues: 
 
 #8
+
 #7
 
 # Release 0.4.1